feat: further page header review updates
