+++ packages/cloud-cognitive/src/components/PageHeader/PageHeader.js
@@ -23,7 +23,7 @@ import {
 } from 'carbon-components-react';
 import { ActionBar } from '../ActionBar/';
 import { BreadcrumbWithOverflow } from '../BreadcrumbWithOverflow';
-import { TagSet } from '../TagSet';
+import { TagSet, string_required_if_more_than_10_tags } from '../TagSet/TagSet';
 import { ButtonSetWithOverflow } from '../ButtonSetWithOverflow';
 import { pkg } from '../../settings';
 import { ChevronUp16 } from '@carbon/icons-react';
@@ -48,6 +48,9 @@ export let PageHeader = React.forwardRef(
       actionBarItems,
       actionBarOverflowAriaLabel,
       actionBarOverflowLabel: deprecated_actionBarOverflowLabel,
+      allTagsModalSearchLabel,
+      allTagsModalSearchPlaceholderText,
+      allTagsModalTitle,
       availableSpace: deprecated_availableSpace,
       background: deprecated_background,
       hasBackgroundAlways,
@@ -60,17 +63,18 @@ export let PageHeader = React.forwardRef(
       collapseHeaderIconDescription,
       collapseHeaderLabel: deprecated_collapseHeaderLabel,
       collapseHeaderToggleWanted: deprecated_collapseHeaderToggleWanted,
+      collapseTitle,
+      disableBreadcrumbScroll,
       expandHeaderIconDescription,
       expandHeaderLabel: deprecated_expandHeaderLabel,
       hasCollapseHeaderToggle,
-      preCollapseTitleRow: deprecated_preCollapseTitleRow,
-      disableBreadcrumbScroll,
-      preventBreadcrumbScroll: deprecated_preventBreadcrumbScroll,
       navigation,
       pageActions,
       pageActionsOverflowLabel,
       pageHeaderOffset,
-      collapseTitle,
+      preCollapseTitleRow: deprecated_preCollapseTitleRow,
+      preventBreadcrumbScroll: deprecated_preventBreadcrumbScroll,
+      showAllTagsLabel,
       subtitle,
       tags,
       title,
@@ -426,14 +430,6 @@ export let PageHeader = React.forwardRef(
       loading: titleLoading,
     } = titleShape;
 
-    const tempTagSetOverflowProps = {
-      // temp fix pending changes to PageHeader implementation
-      allTagsModalTitle: 'All tags',
-      allTagsModalSearchLabel: 'Search all tags',
-      allTagsModalSearchPlaceholderText: 'Search all tags',
-      showAllTagsLabel: 'View all tags',
-    };
-
     useResizeDetector({
       onResize: handleResizeActionBarColumn,
       targetRef: sizingContainerRef,
@@ -488,7 +484,6 @@ export let PageHeader = React.forwardRef(
                         {breadcrumbItems}
                         {title ? (
                           <BreadcrumbItem
-                            href="#"
                             isCurrentPage={true}
                             className={cx([
                               `${blockClass}__breadcrumb-title`,
@@ -497,7 +492,9 @@ export let PageHeader = React.forwardRef(
                                   collapseTitle,
                               },
                             ])}>
-                            {titleLoading ? <SkeletonText /> : titleText}
+                            <span>
+                              {titleLoading ? <SkeletonText /> : titleText}
+                            </span>
                           </BreadcrumbItem>
                         ) : (
                           ''
@@ -657,8 +654,13 @@ export let PageHeader = React.forwardRef(
                     })}>
                     <TagSet
                       overflowAlign="end"
-                      {...tempTagSetOverflowProps}
-                      tags={tags}
+                      {...{
+                        allTagsModalSearchLabel,
+                        allTagsModalSearchPlaceholderText,
+                        allTagsModalTitle,
+                        showAllTagsLabel,
+                        tags,
+                      }}
                     />
                   </Column>
                 </Row>
@@ -686,8 +688,13 @@ export let PageHeader = React.forwardRef(
                     })}>
                     <TagSet
                       overflowAlign="end"
-                      {...tempTagSetOverflowProps}
-                      tags={tags}
+                      {...{
+                        allTagsModalSearchLabel,
+                        allTagsModalSearchPlaceholderText,
+                        allTagsModalTitle,
+                        showAllTagsLabel,
+                        tags,
+                      }}
                     />
                   </Column>
                 ) : null}
@@ -783,6 +790,30 @@ PageHeader.propTypes = {
     PropTypes.string,
     'Property renamed to `actionBarOverflowAriaLabel`.'
   ),
+  /**
+   * When tags are supplied there may not be sufficient space to display all of the tags. This results in an overflow
+   * menu being shown. If in the overflow menu there is still insufficient space this label is used in a dialog showing
+   * all tags.
+   *
+   * **Note: Required if more than 10 tags**
+   */
+  allTagsModalSearchLabel: string_required_if_more_than_10_tags,
+  /**
+   * When tags are supplied there may not be sufficient space to display all of the tags. This results in an overflow
+   * menu being shown. If in the overflow menu there is still insufficient space this placeholder is used in a dialog
+   * showing all tags.
+   *
+   * **Note: Required if more than 10 tags**
+   */
+  allTagsModalSearchPlaceholderText: string_required_if_more_than_10_tags,
+  /**
+   * When tags are supplied there may not be sufficient space to display all of the tags. This results in an overflow
+   * menu being shown. If in the overflow menu there is still insufficient space this title is used in a dialog showing
+   * all tags.
+   *
+   * **Note: Required if more than 10 tags**
+   */
+  allTagsModalTitle: string_required_if_more_than_10_tags,
   // DEPRECATED SEE children
   availableSpace: deprecateProp(
     PropTypes.node,
@@ -939,6 +970,14 @@ PageHeader.propTypes = {
     PropTypes.bool,
     'Prop renamed to `disableBreadcrumbScroll`.'
   ),
+  /**
+   * When tags are supplied there may not be sufficient space to display all of the tags. This results in an overflow
+   * menu being shown. If in the overflow menu there is still insufficient space this label is used to offer a
+   * "View all tags" option.
+   *
+   * **Note: Required if more than 10 tags**
+   */
+  showAllTagsLabel: string_required_if_more_than_10_tags,
   /**
    * Sitting just below the title is this optional subtitle that provides additional context to
    * identify the current page.
