+++ packages/cloud-cognitive/src/components/PageHeader/PageHeaderDemo.data.js
@@ -1,61 +1,5 @@
-//
-// Copyright IBM Corp. 2020, 2020
-//
-// This source code is licensed under the Apache-2.0 license found in the
-// LICENSE file in the root directory of this source tree.
-//
-
-import React from 'react';
-// import { action } from '@storybook/addon-actions';
-
-import {
-  BreadcrumbItem,
-  Column,
-  Grid,
-  Header,
-  HeaderName,
-  Row,
-  Tab,
-  Tabs,
-  Table,
-  TableHead,
-  TableHeader,
-  TableBody,
-  TableRow,
-  TableCell,
-} from 'carbon-components-react';
-import {
-  Printer16,
-  Security24,
-  Settings16,
-  VolumeMute16,
-} from '@carbon/icons-react';
-
-import { pkg, carbon } from '../../settings';
-import { getStorybookPrefix } from '../../../config';
-import { ActionBarItem } from '../ActionBar';
-import { PageHeader } from '.';
-const storybookPrefix = getStorybookPrefix(pkg, PageHeader.displayName);
-
-import styles from './_storybook-styles.scss'; // import index in case more files are added later.
-import mdx from './PageHeader.mdx';
-
-const storyClass = 'page-header-stories';
-
-export default {
-  title: `${storybookPrefix}/${PageHeader.displayName}`,
-  component: PageHeader,
-  subcomponents: {
-    ActionBarItem,
-  },
-  parameters: { styles, layout: 'fullscreen', docs: { page: mdx } },
-  decorators: [
-    (story) => <div className={`${storyClass}__viewport`}>{story()}</div>,
-  ],
-};
-
 /* cSpell:disable */
-const demoTableData = [
+export const demoTableData = [
   {
     Index: 'C8205413-BD94-0B62-DB0A-BE26CF8C60CD',
     Date: '1623141365',
@@ -767,118 +711,5 @@ const demoTableData = [
     Description: 'lorem semper',
   },
 ];
-const demoTableHeaders = Object.keys(demoTableData[0]);
+export const demoTableHeaders = Object.keys(demoTableData[0]);
 /* cSpell:enable */
-
-const TemplateDemo = () => {
-  return (
-    <>
-      <style>{`.${carbon.prefix}--modal { opacity: 0; }`};</style>
-      <div className={`${storyClass}__app`}>
-        <Header aria-label="IBM Platform Name">
-          <HeaderName href="#" prefix="IBM">
-            Cloud Cognitive application
-          </HeaderName>
-        </Header>
-        <div
-          className={`${storyClass}__content-container`}
-          style={{
-            // stylelint-disable-next-line carbon/layout-token-use
-            marginTop: '48px',
-          }}>
-          <PageHeader
-            breadcrumbOverflowAriaLabel="Open and close additional breadcrumb item list."
-            breadcrumbItems={
-              <>
-                <BreadcrumbItem href="../../../homepage">
-                  Homepage
-                </BreadcrumbItem>
-                <BreadcrumbItem href="../../Reports">Reports</BreadcrumbItem>
-                <BreadcrumbItem href="../June2021">June 2021</BreadcrumbItem>
-              </>
-            }
-            actionBarItems={[
-              { key: '1', renderIcon: Printer16, iconDescription: `Print` },
-              { key: '2', renderIcon: Settings16, iconDescription: `Settings` },
-              { key: '3', renderIcon: VolumeMute16, iconDescription: `Mute` },
-            ]}
-            actionBarOverflowAriaLabel="Show more action bar items"
-            title={{
-              text: 'Authentication activity',
-              loading: false,
-              icon: Security24,
-            }}
-            disableBreadcrumbScroll
-            pageHeaderOffset={48} // 48px is the size of the global header. A more elegant way of passing this could be found.
-            pageActions={[
-              {
-                kind: 'secondary',
-                label: 'Acknowledge',
-                onClick: () => {},
-              },
-              {
-                kind: 'primary',
-                label: 'Escalate',
-                onClick: () => {},
-              },
-            ]}
-            subtitle="This report details the monthly authentication failures"
-            navigation={
-              <Tabs>
-                <Tab label="Summary" />
-                <Tab label="Region 1" />
-                <Tab label="Region 2" />
-                <Tab label="Region 3" />
-              </Tabs>
-            }
-            tags={[
-              {
-                type: 'cyan',
-                label: 'Not urgent',
-              },
-              {
-                type: 'red',
-                label: 'Security',
-              },
-            ]}>
-            <p>Severity 1: 0</p>
-            <p>Severity 1: 814</p>
-            <p>Severity 3: 3,108</p>
-          </PageHeader>
-          {
-            <Grid className={`${storyClass}__dummy-content`} narrow={true}>
-              <Row>
-                <Column
-                  sm={4}
-                  md={8}
-                  lg={16}
-                  className={`${storyClass}__dummy-content-block`}>
-                  <Table>
-                    <TableHead>
-                      <TableRow>
-                        {demoTableHeaders.map((header) => (
-                          <TableHeader key={header}>{header}</TableHeader>
-                        ))}
-                      </TableRow>
-                    </TableHead>
-                    <TableBody>
-                      {demoTableData.map((row) => (
-                        <TableRow key={row.Index}>
-                          {Object.keys(row).map((key) => {
-                            return <TableCell key={key}>{row[key]}</TableCell>;
-                          })}
-                        </TableRow>
-                      ))}
-                    </TableBody>
-                  </Table>{' '}
-                </Column>
-              </Row>
-            </Grid>
-          }
-        </div>
-      </div>
-    </>
-  );
-};
-
-export const PageHeaderDemo = TemplateDemo.bind({});
